Move authorization methods to the IIIFResource
More general types of resource do not use auth
